+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- **Breaking:** Add "reset" sub-dictionary to Bullet interface config
 - BulletInterface: Refactor internal reset functions
 - Update mpacklog.cpp to v3.1.0